pyproject.toml
[project]
name = "procmon-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "rich>=14.0.0",
    "python-dotenv>=1.0.0",
    "pynvml>=11.5.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.2"
]

[project.scripts]
procmon = "src.procmon.cli:main"
